Completely switched from eclipse to IntelliJ

+++ src/io/github/seanboyy/lotReset/mca/ReadMCA.java
@@ -1,18 +1,18 @@
 /*
  * This is a part of LotReset, licensed under the MIT License (MIT).
- * 
+ *
  * Copyright (c) 2016 Seanboyy (Sean Bamford)
- * 
+ *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- * 
+ *
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTIBILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -90,7 +90,7 @@ public class ReadMCA{
                     int numSectors = offset & 255;
                     if (sectorNumber + numSectors > open.size()){
                         x[(int) a] = 0;
-                        file.seek((long)(a * 4));
+                        file.seek(a * 4);
                         file.writeInt(x[(int) a]);
                         for (int j = 0; j < numSectors; ++j){
                             open.set(sectorNumber + j, true);
@@ -104,7 +104,7 @@ public class ReadMCA{
                         int length = file.readInt();
                         if (length > 4096 * numSectors){
                             x[(int) a] = 0;
-                            file.seek((long)(a * 4));
+                            file.seek(a * 4);
                             file.writeInt(x[(int) a]);
                             for (int j = 0; j < numSectors; ++j){
                                 open.set(sectorNumber + j, true);
@@ -114,7 +114,7 @@ public class ReadMCA{
                         }
                         else if (length <= 0){
                             x[(int) a] = 0;
-                            file.seek((long)(a * 4));
+                            file.seek(a * 4);
                             file.writeInt(x[(int) a]);
                             for (int j = 0; j < numSectors; ++j){
                                 open.set(sectorNumber + j, true);
@@ -150,7 +150,7 @@ public class ReadMCA{
             }
         }
     }
-    
+
     /**
      * Reads Section data from chunk
      * @param is chunk data stream
@@ -160,23 +160,23 @@ public class ReadMCA{
      * @since 3.0
      */
     public static CompoundTag getSection(DataInputStream is, int Y) throws IOException{
-    	CompoundTag tag = StreamTools.read(is);
-    	CompoundTag emptyTag = new CompoundTag();
-    	CompoundTag level = tag.getCompoundTag("Level");
-    	ListTag sections = level.getListTag("Sections", 10);
-    	CompoundTag section = sections.getCompoundTagAt(Y);
-    	if(!(section.equals(emptyTag))){
-        	byte y = ((ByteTag)section.getTag("Y")).getByte();
-        	if (y == (byte)Y){
-        		return section;
-        	}
-        	else{
-        		return null;
-        	}
-    		
-    	}
-    	else{
-    		return emptyTag;
-    	}
+        CompoundTag tag = StreamTools.read(is);
+        CompoundTag emptyTag = new CompoundTag();
+        CompoundTag level = tag.getCompoundTag("Level");
+        ListTag sections = level.getListTag("Sections", 10);
+        CompoundTag section = sections.getCompoundTagAt(Y);
+        if(!(section.equals(emptyTag))){
+            byte y = ((ByteTag)section.getTag("Y")).getByte();
+            if (y == (byte)Y){
+                return section;
+            }
+            else{
+                return null;
+            }
+
+        }
+        else{
+            return emptyTag;
+        }
     }
 }
